// src/generated/types/activityLog.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */
import type { ActivityLogEntity } from './activityLogEntity';
import type { SiteBasicDto } from './siteBasicDto';
import type { UserDto } from './userDto';

/**
 * Activity log data
 */
export interface ActivityLog {
  id?: string;
  entity?: ActivityLogEntity;
  category?: string;
  action?: string;
  /** Activity Log creation unix timestamp */
  date?: number;
  site?: SiteBasicDto;
  deviceId?: number;
  hostname?: string;
  user?: UserDto;
  details?: string;
  hasStdOut?: boolean;
  hasStdErr?: boolean;
}